Log captured flags to a separate, tailable log
Into "captured_flags.log" at the server's working directory.

The log is structured to be easily parsable by simple regex,
maybe before the competition is up, someone will find a way to
notify anyone who's interested.

Logs to both stdout and that file, simultaneously.

The log filename is configured as a constant at the top of `queries.go`
which is the only place this logger is needed.

This is the most GOLANG I had to write so far. Fun times!