Add native timestamp to Results in DB
While the BSON ObjectID has the timestamp built in,
the _id field cannot be used for any form of date queries.
Doing timeseries analysis of data, as we have wanted to
do before, is impossible without adding a timestamp to
each result, as I have done here.

No queries rely on the new timestamp as of this commit.