checks.go: Add event-wide & single-check toggles
Now, the whole check service can be put into a sleeping state
by setting the "on_break" flag in the config to `true`.

Individual checks can be disabled until they are ready to be
scored with the `disabled = true|false` flag. If no flag is
given, the check will be enabled.

Combined with live-reloading, you could use the system with
scheduled `at` jobs to automate out the whole event schedule.