Remove log dependency for startup errors
At the moment, the only messages that get logged out
by the command initialization, are fatal errors. By
migrating to simple fmt + os.Exit(), this fail-fast
notion becomes a bit more explicit.

More importantly, the 'cmd' and 'server' packages will
not have to be designed to share loggers in the future,
when everything will be centrally logged.