Sub-divide admin/blackteam dashboard
Users with access to a group (or all) challenges
will be able to switch between configuring the ctf,
and getting an breakdown of how all the different
teams are doing.

The separate CTF js has been moved under the
'admin-dashboard.js', which will ALSO includes logic
for getting the score breakdowns.