Add 'Submissions per flag' counts to admin dash
I've implemented these aggregates of event results
as "breakdowns", because it reads better than mongo's
"aggregate" or "pipe results".

This feature was requested by one of the challenge
organizers last time, as it helps determine which
challenges are "too difficult".

There are also stubs for the next breakdown feature,
'Each Teams Captured Flags'.