Add public challenges as config.toml option
Removes the hardcoded "Rusted Bunions" challenges.

With this, any number of challenges can be made "public",
and they will show up on the `Challenges` page if the
user is logged in. If there are no public challenges, the
page no longer appears on the navbar.

Also fixed the Challenges page title display, to properly
list all the challenges using a unique-checking function
(thanks, Mrchief, of stackoverflow:
http://stackoverflow.com/a/18923480)