Add docker-compose support
This helps document all our dependencies, and produces
an artifact that fully runs.

NOTE: Due to the way the app is coded at the moment,
the admin account will be created with a blank password.
A hack to log in requies the browser's developer console:

  `document.getElementsByTagName("form")[0].submit()`

Once logged in, the admin password can be changed from
the dashboard. We planned on changing how this initial
setup is done, so this hack should not be required for
long.