checks_test.go: Swap ping for arp
Ping is unavailable in TravisCI containers due to its
status as a setuid binary:
https://github.com/travis-ci/travis-ci/issues/3080

I investigated further using the travis cli
(https://github.com/travis-ci/travis.rb), and found
that the `arp` command should be a suitable replacement
for this test.

I dislike the idea of bending over to accomodate our CI
environment, but this may be a sign that this test is too
fragile and should be redone later.