Add score breakdown to web display
The big scoreboard chart is now a stacked bar graph,
representing points in the CTF and Services categories.
Tooltip, label, and legend styles were tweaked to improve
legibility and responsiveness.

Also, fixed the spaces between scores where a comma would be.
This spacing caused the point labels to look really 'off'.

The table below the chart has been updated to include all
of this same information.