Add Postgres support for scoring
> Two databases?
Yeah, I'm sorry.

This commit adds the option of duplicating all competition results
into Postgres, enabling data exploration with other tools that
prefer timeseries query support. The target here is Grafana,
which let's us show off dozens of fancy graphs that can be made
rapidly with ease, and have interactive actions that even the
average joe can find useful.

The idea is to be able to better explain what goes on with regards
to the scoring throughout the competition. This was too difficult
with Mongo due to the crippled query language it is built around.