Add baseline postgres schema migration, no changes.
Documentation is inlined in this first sql script.

I'm going to step back from this and fix other things, so
I wanted comments to remind myself with, in the future.

Also because data is the heart of every application,
so having an explanation for its design is critical.

The comments can later be massaged and made into a low-level document.