Queries: Fixed `null` response from several routes
The mongo driver in golang will allocate arrays and structs for you,
but not if there is no data in the results of a query. If you feed
the query methods a nil slice, it will leave it nil. If you feed
them an empty slice, you'll have that empty slice.

nil marshalled to JSON becomes `null`
Empty slices will become `[]`

In the case of an empty database, the admin panel would fail to load
completely because several routes would return a literal `null` string,
which is totally valid JSON but not iterable like we were expecting.

Now, an empty database will correctly feed back empty arrays, which
can be `iterated over`. The admin panel no longer sort of explodes.