Auth: Switch from gorilla/sessions to alexedwards/scs
Our auth needed to be gone over anyway, and while I was investigating,
I discovered a newer library for handling session data.

Advantages of scs:
+ Modern crypto that combines encryption & authentication.
    - See: https://godoc.org/golang.org/x/crypto/nacl/secretbox
    - gorilla's project lead is planning to add it as an option in
        securecookie, so its definitely good crypto.
+ Nicer API, resulting in simpler code.
    It is based around `context` from go 1.7, which gives it an
    edge in interop over gorilla's own context package.

I crawled over the scs codebase, and it looks solid. I wanted to be sure
before making the switch, though.

Of course, if gorilla ever releases v2 of sessions / securecookie,
I could see switching back. That would be purely as a gesture of
support for elithrar, the ever hard-working golang enthusiast who
leads development of the gorilla toolkit.