Auth: Persist session signing secret to DB
Fixes a long standing bug where restarting the server would invalidate
everyone's sessions. With the way we were using gorilla/sessions, users
would have to manually delete their cookies just to be able to log in
again, which was a huge headache! Not to mention the disruption of
service. Or the hundreds of log ins I've had to do while developing.

Not sure why I hadn't done this sooner.