Auth: Refactor a little and add godocs
I moved the CheckSessionID middleware out of middleware.go and
put it with the rest of the authentication stuff. Personally, it
feels more conherent to have it all together, and I think the fact
that it's related to logins is more relevant than the fact that it's
a middleware function.

Either way, I added godocs to make everything clear and eaiser to
maintain.