Auth: Cookies last 1 week, option to share over http
The event typically lasts a weekend, so a week-long session should
be sufficient. No need to worry about updating the expiration.

The session manager can be instructed to allow cookies over http,
which would usually be unsafe, but can be handy for testing.
Also, I noticed that the secure flag was not being set before the
switch to scs, even though gorilla/sessions supports it. Oops.