Generate first draft model files for postgres
I used github.com/knq/xo, modified the templates used to generate the
golang model code for interacting with the postgres tables. That was
just a jumping off point, so I'm not committing the customized
templates, and the generated code I then updated further by hand
anyways.

There's not a lot of models, so I prefer this approach to using a
query builder or an ORM. Plus, the scoring aggregations will
require custom SQL anyway, so why waste time figuring out an ORM
and all its baggage.

The scoring queries aren't in here yet.
The models package will be moved in the future, as well.