Port batch insert queries
Pgx supports the binary copy protocol for Postgres, which I used for
batch inserting service monitor status updates, but not for other
insert many ops. It's a bit of a pain, boilerplate-wise to support,
feels brittle with the requirement to stick values in interface{}
slices, and using a simple SQL transaction is much easier to follow.