Use serializable isolation level when scoring flags
I remembered this blog post:
https://shiroyasha.io/transaction-isolation-levels-in-postgresql.html

So by default, Postgres transactions will not protect against
all kinds of data races. Specifying the isolation level as such
should be enough to prevent contestants for doing some form of
double-submit on a challenge, letting them sneak in twice as many
points (or more).