Add first tests for postgres
Tests are simple sanity checks that show basic querying works.

This uses a library that relies on a real postgres db with real data.
No mocks or anything, which I feel is better as the tests become
more true to reality. Now the problem is all the bending over backwards
to get the test scaffolding to work flawlessly. This will not likely
work well with travisCI, either, as I would need to pull custom
extension modules and have the provided postgres instance load them,
with the correct db user, etc.

One problem with the library is that it seems to run off and try
to toggle too many triggers, which causes the tests to fail unless
they are run as a db superuser. I would prefer the tests get run
as the same user the application runs as, so I'll look for a
workaround.