Add tests for scoring, issue with decimal math
First, test data was adjusted because I am assuming that if an admin
goes to disable a service/ctf challenge mid-competition, they would NOT
want the scoreboard to thrash everyone down as a result. If a
challenge/service is completely broken, deleting it will be enough to
wipe out whatever gains a contestant may have already got, whereas
hiding either of them is just a temporary measure while they get fixed
in some way.

The scoring tests check if scores are accumulated correctly (no brainer!)
But due to the way floating point numbers work under the hood, this can
result in crazy trailing decimals on some scores (5.5 + 2.2 = 7.999994).
The sums are consistent across platforms (per IEEE specification), but
I'm still deciding how to handle this. The numbers could be rounded,
cast to integers, returned as text then parsed as decimal numbers, left
alone and simply have the UI deal with it, etc. I'll figure something
out.