Round calculated scores, don't deal with decimals
After mulling it over, I'm deciding to just round score values.
The exact amount down to the decimal is unimportant. If scores are ever
within a point or two of one another, everyone from contestants to
organizers will be amazed by each team.

Plus, I don't want to deal with decimals in the back end, and on the
front when we go to display them. It will look cleaner to show a whole
number.