Update most template controllers for PG models
This commit mostly involves shuffling variables around and fixing up
broken code. Many model types are cleaner and make for easier to follow
code, which is nice.

I've also moved all the methods not related directly to the view out of
the templates, and instead pass in just the necessary types with a map
called Data. This makes it much easier to trace what data is required
for any given template to render.

I haven't finished updating the admin dashboard, but, as with many admin
dashboard projects, it is very involved. I'm planning to split the admin
dashboard into multiple pages, instead of one monolith like it is now.