Impl more admin-specific api handlers
Using go-chi/render for reading requests/shipping responses is exactly
what I was looking for months ago. It seriously cuts down on the
boilerplate.

With that, I'm cleaning up duplicative code as I go where I can.
The result is that many of the handlers are just two function calls with
error handling. I think with this foundation, it will be easy to add all
the service-configuration routes, when I get to that.