Make test scaffold reusable; fix authentication tests
I pulled the test setup code out of server/models, and made it it's
own little package that gets loaded by tests. With a few tweaks, now
it's easy to use the same batch of convenient globals to test multiple
parts of the program. I planned on splitting the app up a bit more, so
this will definitely come in handy.

Also I updated the authentication tests, and test data, to work with
the new postgres stuff, so now we can rest assured that at least logins
won't break unexpectedly.

Though, I am running into a very bizarre bug where `go test` is not
properly exiting with a non-zero status code when there are test
failures. I can't find any open issues about it, so I'm not sure what
can be done.