Migrate check service to new models; only one config.toml now
The service checker now is encapsulated into the Monitor type (as
in "service monitor"). Documentation forthcoming - as it always is.

I had to think over the lifecycles of the check/monitor service
several times to find a strategy that keeps it's state up to date
with the database and event timeline. I expect there to be bugs.
Concurrency is tricky.

In addition, I whittled down the config files to just one, as there
are only a few sections unique to web server vs the service monitor,
and it's just easier to not have to juggle the two with all their
duplication.
The config file now specifies when set breaks should occur during
an event, which is necessary to calculate scores per service check
accurately. I may also add a way to arbitrarily put the event on
hold, but that will make it much harder to accurately score teams
on their services.

Still need to make some updates to postgres to create a trigger
on insert/updates into teams/service, to NOTIFY listeners of the
update.