Automatic starting of individual services; .sql trigger setup
I forgot the one remaining aspect of the service configuration, which is
that it should be possible to automate it by simply setting the times
each service starts. In practice, we tend to play things by ear, but
I still like the idea of 'set it up and let it run' for the whole event.

Also add the SQL necessary in the setup .sql script for the triggers
that the service monitor relies on.

This whole week, I've had the feeling that it would have been completely
fine to simply fetch all teams + services with the appropriate WHERE
clauses before every run of the service monitor. It would centralize all
the concurrency stuff on Postgres, remove a ton of burden and potential
bugs from the application code, and be easier to reason about.
But then, it might be a bit less fast.
Maybe I made a mistake, and overcomplicated the implementation here.