Add api for managing files for CTF challenges; files go in `data/`
Originally, I designed these to be tracked by the database under the
challenge_file table, but as I delved into how to make the download file
names come out right[1][2] to the client, I decided it wouldn't be worth
it. As a plus, I saw how easy it would be to reuse the service scripts
functions for managing CTF files.

To keep track of all these runtime generated files, I've moved the
default fs location for logs, scripts, and now ctf files into the
`data/` directory.

[1] https://tools.ietf.org/html/rfc2183
[2] https://stackoverflow.com/q/93551