Services page refactor; Fix pesky service status aggregate ordering
With my sample data, the order that the api endpoint used on the
services page returned was based on db insert order of each row. For
unit tests and during the actual events, db insert order is always
already 'sorted' by team_id. The sample data dumped from mongo did not
have that quality, so all of the sudden the page was subtly broken. By
using a sub-select to sort the nested query in scoringservices.go, the
ordering is more stable. Sorting pre-sorted data in postgres is
fast, so performance change is negligible for the usual case.