Automatically sync service uptime display on clients
The service uptime matrix (all the green/red/yellow circles on the
scoreboard page) now handle when services are added/removed,
automatically.

Uses a bunch of jquery and a hand-written, flimsy DOM diffing algo,
that every Js 201X framework does internally.

If the number of teams changes, the Js running should detect this and it
will just reload the whole page.

Service add/delete can happen so suddenly that contestants might not
notice, but if they're given a warning through other comm channels,
they should be able to expect the change.