Separate 'public' and 'hidden' challenges, as originally intended
Somewhere in the design, I must have had my brain skip a beat. I
modelled challenges so they could be public or anonymous, and then when
I went to implement them however many weeks later, I mistook my design
to mean 'hidden' == 'disabled'. But I don't see the use in disabling a
challenge. If it's bad, just delete it, and the scoreboard will pick up
the changes.

Now, a non-hidden/named challenge is required to be guessed with it's
name and vice-versa for anonymous challenges.