Begin redoing staff pages (config ctf, teams, etc.)
I'm segmenting the dashboard into distinct, focused pages, instead of
the monolith monster it had started to become. Started by redoing CTF,
which has all the markup in place.

Noticed long ago that not all api endpoints sort their data in the SQL
query, so that will be updated during the rework.