ctf config: Add button for creating just one challenge at a time
By request.
Adding one at a time makes more sense now. I had to hack in an endpoint
in the ctf routes namespace, but I'll rearrange that later to be like
the rest of the resource routes.

I also plan to look into adding files as the challenge is being created.
I don't think the UI supports it very cleanly right now.