UI: Update to bootstrap 4 & jquery 3; Move libs
On the bleeding edge of last decade's ui design!

Flexbox is great, but it sure was a hassle recovering the subtle
'freebies' from bs3 in bs4. Basically a game of hunt and peck, as the
migration guides don't cover everything.

While I was at it, I migrated all the 3rd party css/js into the
lib folder, where it's easier to track.