Ctf ui: Add columns for first team to solve; total solves
Using the timescaledb `first` function, I was able to easily find the
quickest team to solve each challenge (or nobody!). The query for this
isn't the fastest, and should be rewritten at some point.

The CTF dashboard now includes the above, lists every challenge in the
'solves per challenge' table even if they have 0 solves, and shows the
total solves for the entire competition.

Additionally if a team has no flags captured, their tab in the dashboard
is marked, which provides a visual check of which teams may be worth
checking out earlier on in an event.

Also fixed tests.
It's late.