Protect cookies with SameSite "Lax"
A relatively new standard (the RFC is from 2016), SameSite prevents
cookies from being sent in cross-domain contexts. So it's harder for
someone to jack sessions maliciously.

"Lax" is the suggested value for most cases. "Strict" is unnecessary
here.

Continued thanks to alexedwards/scs for support of his session
management library 👏