Update readme for postgres and web ui config
Huge portions of the readme were updated or newly written. It's starting
to get large enough that I feel like breaking it apart. Some pieces
should also be copied into the Web UI, which more staff take the time to
interact with. And in reverse, screenshots of the web ui should be added
to the readme.

I also added in my instructions on running the DB with docker, which I
will flesh out more in the future. Running PG with Docker is becoming
more likely for me, as I'm getting tired of accidentally fighting
database version updates every year.