CTF: Escape ctf files in URLs, and sanitize upload names
Last event, the CTF team wanted to use a file just named `?`, which
revealed a rookie mistake in the API. The 'CMS' is perfectly capable of
saving off files with odd names, but the URLs that get generated to
snatch them later weren't getting url-encoded.

After looking at this for an extended time, what I've realized is that
the CTF / Script API is simply not compatible with the RFCs for URLs.
So, for instance, files with `!` and `@` in them still can't be read
after upload, even though they are valid in filenames on the filesystem.
Any part of the URL path has some set of reserved characters that can't
be transferred as is to the server, which is leaving us with odd edge
cases like these. To fix this completely, the API should have to be
redesigned.

While adjusting that, wondering whether I should url-encode on the
server or the client, I noticed that the API was blindly accepting
filename given on the create endpoint. This was likely exposing a vuln
that could overwrite arbitrary files on the server. It probably didn't
get noticed before because the API for creating CTF files is supposed
to be for trusted users, but there's no need to leave it vulnerable.
See:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition#Directives