[WIP] Add ws API endpoint to watch a log file
Trying out a new websocket library that has a tighter design than
gorilla's trusty websocket lib. Unfortunately, it looks like this
new lib doesn't have "prepared messages", which seriously help with
performance of broadcast to hundreds of clients. For this simple
admin-only endpoint, that's ok, but not for the main scoring page.

Still TODO: add the web ui for displaying the log files.