Blue Team dashboard: Add cnyhackathon-specific tips, del hidden flag
I need to rework the team's dashboards at some point, but we haven't had
the CTF run the old way in a year, where there was a mix of scavenger
hunt & jeopardy styles. Since we don't do hidden input, I've just axed
that part and replaced it with tips for the CNY hackathon event.

There should be some web ui to change these kinds of pages to customize
the event, but it's not a lot different right now from editing any other
template language. Though, a server restart is required, which there
could be ways to avoid that can be looked into later.