Admins: Show event breaks (maybe everyone should see this?); minor fix
Now admins can see a clean display of when the system goes on break,
from the Service checks config page.

This might be the kind of thing that should be shown to everyone.
Or just remember to alert everyone 5-30 minutes before the break starts.

Happened to notice an uneeded query that happens when admins go to their
dashboard. It was easy to clean up, but it's the first instance of that
team-based template-generation being done. Probably for good reason,
because it's adding more complicated interwoven variables to track
across system boundaries.