Admins: Add API to get event config, & set default for service start
I'm tying more things into the event configuration, so it's useful to
expose that data over the API to staff members.
The `/api/staff/event_config` route provides start & stop times, breaks,
and the time interval between checks (which is then further randomized).

When setting up services, it helps to have the date & time inputs
pre-populated to the event's start time. This is when we start most
services.